backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]>=0.34.0
websockets==14.1
httpx>=0.28.1
google-generativeai==0.8.3
google-cloud-firestore==2.19.0
google-adk==1.0.0
elevenlabs==1.9.0
firecrawl-py==1.5.0
playwright==1.49.0
python-multipart==0.0.12
pydantic==2.10.3
python-dotenv==1.0.1
